pyproject.toml
[project]
name = "medgemma-impact-challenge"
version = "0.1.0"
description = "Kaggle MedGemma Impact Challenge"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.1",
    "google-adk[extensions]>=1.23.0",
    "gradio>=6.5.1",
    "langfuse>=3.12.1",
    "marker-pdf>=1.10.2",
    "openinference-instrumentation-google-adk>=0.1.8",
    "pydantic>=2.12.5",
]

[dependency-groups]
dev = ["pytest", "pytest-cov", "ruff", "pre-commit"]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py312"
